chore: remove perpetual agent stubs and rename trpg files (-118 lines)
The perpetual agent system was removed but empty stubs and types
remained in the dashboard API response and TypeScript code.

Removes:
- perpetual_dashboard_json() empty stub from dashboard_http_mdal.ml
- "perpetual" key from dashboard API JSON response
- PerpetualStatus type and always-null signal from dashboard store

Renames (content is now solely MDAL/parse utilities, not TRPG):
- types/trpg.ts → types/mdal.ts (contains only MDAL types)
- api/trpg.ts → api/parse-utils.ts (contains only parse utilities)
- Updates all import paths (3 files) and barrel re-exports (2 files)

Verified: dune build + npx tsc --noEmit pass with 0 errors.
