+++ dashboard/src/api.ts
@@ -6,6 +6,6 @@ export * from './api/mcp'
 export * from './api/keeper'
 export * from './api/dashboard'
 export * from './api/board'
-export * from './api/trpg'
+export * from './api/parse-utils'
 export * from './api/actions'
 export * from './api/autoresearch'
