+++ dashboard/src/types.ts
@@ -2,7 +2,7 @@
 // Domain-specific types are organized in src/types/ subdirectory
 
 export * from './types/core'
-export * from './types/trpg'
+export * from './types/mdal'
 export * from './types/governance'
 export * from './types/dashboard-execution'
 export * from './types/dashboard-mission'
